requirements.txt
pyscard==1.9.7
pyserial==3.4
requests==2.20.1
